Warn users about iOS native libraries in remote debug
Adds detection and UI warnings for iOS native libraries (.a, .dylib, .xcframework, .framework) that are not supported during remote debugging. The plugin now displays a warning panel in the UI and shows a dialog (with a 'Don't Show Again' option) if such libraries are found, and also pushes warnings in the export platform logic. This helps users understand why certain native features may not work during remote debugging sessions.